+++ pyrax/utils.py
@@ -16,7 +16,6 @@ import threading
 import time
 import types
 
-import prettytable
 try:
     import pudb
 except ImportError:
@@ -379,6 +378,20 @@ def get_id(id_or_obj):
         return id_or_obj.id
     except AttributeError:
         return id_or_obj
+
+
+def get_name(name_or_obj):
+    """
+    Returns the 'name' attribute of 'name_or_obj' if present; if not,
+    returns 'name_or_obj'.
+    """
+    if isinstance(name_or_obj, basestring):
+        # It's a name
+        return name_or_obj
+    try:
+        return name_or_obj.name
+    except AttributeError:
+        raise exc.MissingName(name_or_obj)
 
 
 def match_pattern(nm, patterns):